Hacky fix
For some reason it is loading Spree preferences here where
the keys in the styles hash for paperclip use strings, not
symbols (e.g. styles: {"mini" => "48x48>", ...}).  However
paperclip is expecting these to be symbols. This is already
set here, so I change it to ||=, and also just in case
convert the strings to symbols.